Sửa cấu trúc project

+++ src/controller/GameVector.java
@@ -6,10 +6,10 @@ import java.util.DoubleSummaryStatistics;
  * Created by apple on 12/24/16.
  */
 public class GameVector {
-    public int dx;
-    public int dy;
+    public double dx;
+    public double dy;
 
-    public GameVector(int dx, int dy) {
+    public GameVector(double dx, double dy) {
         this.dx = dx;
         this.dy = dy;
     }
